# Conflicts:
#	Makefile
#	boards/LINUX.py
#	libs/graphics/jswrap_graphics.c
#	src/jsvar.h

+++ libs/graphics/bitmap_font_6x8.h
@@ -15,5 +15,5 @@
 #include "graphics.h"
 
 #ifdef USE_FONT_6X8
-void graphicsDrawChar6x8(JsGraphics *gfx, short x1, short y1, char ch, unsigned short size, bool solidBackground);
+void graphicsDrawChar6x8(JsGraphics *gfx, int x1, int y1, char ch, unsigned short size, bool solidBackground);
 #endif
